+++ src/models/jwt.ts
@@ -1,0 +1,6 @@
+import { BaseModel } from "./base_model.js";
+import { Jwt as JwtInterface } from "../interfaces/jwt.js";
+
+export class Jwt extends BaseModel implements JwtInterface {
+  declare jwt: string;
+}